% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineWaterUse.R
\name{combineWaterUse}
\alias{combineWaterUse}
\title{Combine water use data for urban system into functional flows}
\usage{
combineWaterUse(start, end, wu)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{wu}{list of dataframes output by getWaterUse}
}
\value{
wu_flows list of xts objects including:
       surface water withdrawals for theromelectric power (sw_therm),
       surface water withdrawals for potable use (sw_pot),
       surface water withdrawals for nonpotable use (sw_npot),
       groundwater withdrawls for thermoelectric power (gw_therm),
       groundwater withdrawals for potable use (gw_pot),
       groundwater withdrawals for nonpotable use (gw_npot)
}
\description{
This function takes county-level USGS water use data gathered by getWaterUse
and aggregates them for whole urban system according to flows in 
CityWaterBalance
}
\examples{
wu = getWaterUse(c("IL"),c("Cook","Lake"))
wu_flows = combineWaterUse("2000-01-01","2015-01-01",wu)
}

