% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb_scenarios.R
\name{cwb_scenarios}
\alias{cwb_scenarios}
\title{Summarize City Water Balance model output for one or more scenarios}
\usage{
cwb_scenarios(data, p, params = NULL)
}
\arguments{
\item{data}{for \code{CityWaterBalance}}

\item{p}{parameters for \code{CityWaterBalance}}

\item{params}{optional set of additional parameters (e.g., dataframe)}
}
\value{
dataframe of modeled flows, summarized for each scenario
}
\description{
This function runs \code{CityWaterBalance} for one or more sets of parameters 
(i.e., scenarios) and summarizes the output by summing all flows over modeled 
time period and normalizing them to precipitation.
}
\examples{
p <- list("openwat"=0.02,"interc"=0,"et_amp" = 1,"flow_amp" = 1,"run_amp"=3.378,
         "run_css"=0.35, "baseflow_amp" = 1, "nonrev"=0.08,"powevap"=0.012,
         "wastgen"=0.85,"potatm"=0.13,"potinfilt"=0,"npotinfilt"=0.5,
         "evslud"=0,"css_leak"=0.05,"deepgw"=0.5,"dgwloss"=1)
out <- cwb_scenarios(cwb_data,p)
}
