% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWaterImports.R
\name{estimateWaterImports}
\alias{estimateWaterImports}
\title{Estimate water imports}
\usage{
estimateWaterImports(start, end, wu, sw_uses = NULL)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{wu}{list of dataframes, one for each use category, from getWaterUse}

\item{sw_uses}{list of surface water use categories to ignore}

\item{gw_uses}{list of groundwater use categories to ignore}
}
\value{
xts object of imports (monthly)
}
\description{
This function estimates water imports as a component of USGS water use data,
based on source (surface or groundwater) and selected categories of use for 
each
}
\examples{
wu = getWaterUse(c("IL"),c("Cook","DeKalb"))
imports = estimateWaterImports("2000-01-01","2015-12-31",wu,sw_uses=c("Mining","Thermoelectric"))
}

