% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtmoFlows.R
\name{getAtmoFlows}
\alias{getAtmoFlows}
\title{Gather and consolidate all data time series for water balance}
\usage{
getAtmoFlows(start, end, geometry, att, val = NA, latitude)
}
\arguments{
\item{start}{start date in format 'YYYY-MM-DD'}

\item{end}{end date in format 'YYYY-MM-DD'}

\item{geometry}{name of geometry as displayed in GDP}

\item{att}{attribute of geometry as displayed in GDP}

\item{val}{values of attribute as displayed in GDP}

\item{latitude}{(degrees)}
}
\value{
data as xts object
}
\description{
This function gathers precipitation and evapotranspiration data, merges these
with streamflow and water use data, calculates the time-varying balance.  
NOTE:  Shapefile for geometry must already be uploaded to the USGS 
Geo Data Portal (GDP)
}
\examples{
atm <- getAtmoFlows('2010-01-01','2010-12-31','sample:Counties','STATE','RI',
41.824)
}
