% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Gather and consolidate all data time series for water balance}
\usage{
getData(start, end, area, geometry, attribute, flowin, flowout,
  imports = NULL)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{geometry}{name of geometry as displayed in Geo Data Portal (GDP)}

\item{attribute}{name of geometry attribute as displayed in GDP}

\item{flowin}{streamflow into the area (in cubic km)}

\item{flowout}{streamflow out of the area (in cubic km)}

\item{imports}{imports of water into area (in cubic km)}
}
\value{
data with water balance as xts object
}
\description{
This function gathers precipitation and evapotranspiration data, merges these
with streamflow and water use data, calculates the time-varying balance
}

