% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvapotranspiration.R
\name{getEvapotranspiration}
\alias{getEvapotranspiration}
\title{Assemble an evapotranspiration times series}
\usage{
getEvapotranspiration(start, end, geometry, attribute)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{geometry}{"name" of area of interest (AOI) in the GDP}

\item{attribute}{"attribute" of AOI to aggegate over}
}
\value{
ET
}
\description{
This function aggregates monthly SSEBop output for a given area using the 
USGS Geo Data Portal (GDP)
}
\examples{
et = getEvapotranspiration("2010-09-01","2010-12-31","sample:CONUS_states","STATES") 
}

