% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreamflow.R
\name{getStreamflow}
\alias{getStreamflow}
\title{Gather time series of streamflow data}
\usage{
getStreamflow(start, end, gages, plot = NULL)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{gages}{list of USGS gage numbers}

\item{plot}{option to plot the data}
}
\value{
list list of 3: sites, site numbers, flow data as xts object
}
\description{
This function gathers daily average streamgage data for a group of gages
from USGS NWIS
}
\examples{
flow = getStreamflow("2000-01-01","2010-12-31",c("05552500","05543500"),1)
}

