% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaterUse.R
\name{getWaterUse}
\alias{getWaterUse}
\title{Gather time series of water use data}
\usage{
getWaterUse(states, counties, years = "ALL", convert = NULL)
}
\arguments{
\item{states}{list of state abbreviations}

\item{counties}{list of county name lists for each state}

\item{years}{list of years of available data}

\item{convert}{conversion factor from MGD to user choice}
}
\value{
list of dataframes, one for each water source and quality
}
\description{
This function gathers and summarizes water use data from USGS NWIS by county,  
year, source (surface or groundwater) and quality (fresh or saline)
}
\examples{
wu = getWaterUse(c("IL"),c("Cook","DeKalb"))
}

