% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge data sources into input for CityWaterBalance}
\usage{
mergeData(area, pet, inflow, outflow, sfpart, wu, ws_imports = NULL,
  etc_imports = NULL, wweff = NULL)
}
\arguments{
\item{streamflow}{xts object, e.g. output of combineStreamflow}

\item{wastewater}{xts object of wastewater flows}
}
\value{
xts of total flow for each timestep
}
\description{
This functions converts units and merges data needed by CityWaterBalance
}
\examples{
gages = c("05551540","05552500")
flow = getStreamflow("2000-01-01","2010-12-31",gages)
flow = gapfillStreamflow(flow,list(c(gages[1],gages[2])))
flow = combineStreamflow(flow,c(0.5,0.5))

}

