% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGlobalBalance.R
\name{plotGlobalBalance}
\alias{plotGlobalBalance}
\title{Plot urban water balance}
\usage{
plotGlobalBalance(data, annual = NULL)
}
\arguments{
\item{data}{xts object created by getData}

\item{annual}{flag indicating whether to plot annual averages}
}
\value{
plot
}
\description{
This function plots output from getData, showing time-variant water balance
at monthly or annual-aggregated resolution.
}

