% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStreamflow.R
\name{plotStreamflow}
\alias{plotStreamflow}
\title{Plot discharge at a set of gages}
\usage{
plotStreamflow(flowlist)
}
\arguments{
\item{flowlist}{list object created by getStreamflow}
}
\value{
plot
}
\description{
This function plots output from getStreamflow
}
