% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityWaterBalance.R
\name{CityWaterBalance}
\alias{CityWaterBalance}
\title{Tracks flows and storages of water internal to an urban system}
\usage{
CityWaterBalance(data, fixed_pms, var_pms)
}
\arguments{
\item{data}{xts or zoo object with date index and columns of data for:
precipitation (prcp),
evapotranspiration (et),
streamflow in (inflow),
streamflow out (outflow),
water supply imports (ws_imports),
other imports (etc_imports),
surface water withdrawals for thermoelectric power (sw_therm),
surface water withdrawals for potable use (sw_pot),
surface water withdrawals for nonpotable use (sw_npot),
groundwater withdrawls for thermoelectric power (gw_therm),
groundwater withdrawals for potable use (gw_pot),
groundwater withdrawals for nonpotable use (gw_npot),
deep groundwater recharge (dgr),
combined sewer overflow events (cso),
wastewater treatment plant effluent (wtpe)}

\item{fixed_pms}{list of parameter values for:
fraction of area that is open water (openwat),
fraction of et from interception (interc)
fraction of runoff diverted to css (run_css),
fraction of potable water supply lost to leaks (nonrev),
fraction of cooling water that evaporates (powevap),
fraction of potable use that returns to css (wastgen),
fraction of potable use that evaporates (potatm),
fraction of potable use that infiltrates (potinfilt),
fraction of nonpotable use that infiltrates (npotinfilt),
fraction of wastewater that evaporated from sludge (evslud),
fraction of wastewater effluent from I&I (css_leak)}

\item{var_pms}{list of time-variant parameter values for:
runoff, as a fraction of precip (roff)}
}
\value{
list of dataframes for 1) global flows, 2) internal flows, 
         3) storages, 4) global balance, 5) internal balance
balance, flows and 2) storages
}
\description{
This function tracks the environmental and manmade flows gathered using other
CityWaterBalance functions as they move through pathways and storages
within the urban system.
}

