% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityWaterBalance.R
\name{CityWaterBalance}
\alias{CityWaterBalance}
\title{Tracks flows and storages of water internal to an urban system}
\usage{
CityWaterBalance(data, p, print = TRUE)
}
\arguments{
\item{data}{xts or zoo object with date index and columns of data for:
precipitation (prcp),
evapotranspiration (et),
streamflow in (inflow),
streamflow out (outflow),
water supply imports (ws_imports),
other imports (etc_imports),
surface water withdrawals for thermoelectric power (sw_therm),
surface water withdrawals for potable use (sw_pot),
surface water withdrawals for nonpotable use (sw_npot),
groundwater withdrawls for thermoelectric power (gw_therm),
groundwater withdrawals for potable use (gw_pot),
groundwater withdrawals for nonpotable use (gw_npot),
deep groundwater recharge (dgr),
combined sewer overflow events (cso),
wastewater treatment plant effluent (wtpe)
runoff estimate (runoff)
baseflow estimate (baseflow)}

\item{p}{list of fixed parameter values for:
fraction of area that is open water (openwat),
fraction of pet lost to interception (interc),
amplification factor for et (et_amp),
amplification factor for outflow (flow_amp),
amplification factor for runoff (runamp),
amplification factor for baseflow (baseflow_amp),
fraction of runoff diverted to css (run_css),
fraction of potable water supply lost to leaks (nonrev),
fraction of cooling water that evaporates (powevap),
fraction of potable use that returns to css (wastgen),
fraction of potable use that evaporates (potatm),
fraction of nonpotable use that infiltrates (npotinfilt),
fraction of wastewater that evaporated from sludge (evslud),
fraction of wastewater effluent from gw infiltration (css_leak)
fraction of groundwater from deep, confined aquifers (deepgw)
fraction of deep groundwater not replaced by inflow (dgwloss)}

\item{print}{option to print messages}
}
\value{
list of dataframes for 1) global flows, 2) internal flows, 
       3) storages, 4) global balance, 5) internal balance and 2) storages
}
\description{
This function tracks the environmental and manmade flows gathered using other
CityWaterBalance functions as they move through pathways and storages
within the urban system.
}
