% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityWaterBalance.R
\name{CityWaterBalance}
\alias{CityWaterBalance}
\title{Tracks flows of water through the urban system}
\usage{
CityWaterBalance(data, p, print = TRUE)
}
\arguments{
\item{data}{xts or zoo object with date index and columns of data for:\cr
precipitation (prcp) \cr
evapotranspiration (et) \cr
streamflow in (inflow) \cr
streamflow out (outflow) \cr
water supply imports (ws_imports) \cr
other imports (etc_imports) \cr
surface water for industrial uses  (sw_ind) \cr
surface water withdrawals for potable use (sw_pot) \cr
surface water withdrawals for nonpotable use (sw_npot) \cr
groundwater withdrawals for industrial uses (gw_ind) \cr
groundwater withdrawals for potable use (gw_pot) \cr
groundwater withdrawals for nonpotable use (gw_npot) \cr
deep groundwater recharge (dgr) \cr
wastewater treatment plant effluent (wtpe) \cr
baseflow estimate (baseflow) \cr}

\item{p}{list of fixed parameter values for: \cr
multiplier for prcp (prcp_mult) \cr
multiplier for et (et_mult) \cr
multiplier for outflow (flow_mult) \cr
multiplier for baseflow (bf_mult) \cr
multiplier for deep pumping replaced by lateral flow (dgw_rep) \cr
fraction of study area that is impervious (imperv) \cr
fraction of pet lost to interception (interc) \cr
fraction of area that is open water (open_wat) \cr
fraction of runoff diverted to combined sewer system (run_css) \cr
fraction of runoff diverted to separated sewer system (run_sss) \cr
fraction of potable water supply lost to leaks (nonrev) \cr
fraction of industrial water that evaporates (ind_evap) \cr
fraction of potable use that returns to sewers (wast_gen) \cr
fraction of potable use that evaporates (pot_atm) \cr
fraction of nonpotable use that infiltrates (npot_infilt) \cr
fraction of wastewater that evaporates from sludge (slud_evap) \cr
fraction of wastewater effluent from gw infiltration (leak_css) \cr
fraction of separated sewer effluent from gw infiltration (leak_sss) \cr
fraction of groundwater from deep, confined aquifers (dgw) \cr}

\item{print}{option to print messages}
}
\value{
list of dataframes:
 \item{all_flows}{all flows} 
 \item{state_vars}{state variables}
 \item{global_balance}{global water balance} 
 \item{internal_balance}{internal water balance}
}
\description{
This function tracks flows of water as they move through pathways and 
storages within the urban system. Data can be in any self-consistent units.
}
\examples{
p <- list("prcp_mult" = 1, "et_mult" = 1, "flow_mult" = 1, "bf_mult" = 1, "dgw_rep" = 0.5,
         "imperv" = 0.2, "interc" = 0, "open_wat" = 0.02, "run_css" = 0.3, "run_sss" = 0.05,
          "nonrev" = 0.08, "ind_evap" = 0.012, "wast_gen" = 0.85, "pot_atm" = 0.13, "npot_infilt" = 0.5, 
         "slud_evap" = 0, "leak_css" = 0.05, "leak_sss" = 0.05, "dgw" = 0.5)
m <- CityWaterBalance(cwb_data, p) 
}
