% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityWaterBalance.R
\name{CityWaterBalance}
\alias{CityWaterBalance}
\title{Tracks flows and storages of water internal to an urban system}
\usage{
CityWaterBalance(data)
}
\arguments{
\item{data}{xts or zoo object with date index and columns of data for:
precipitation (prcp),
evapotranspiration (et),
streamflow in (inflow),
streamflow out (outflow),
baseflow (bflow),
stormflow (sflow),
water supply imports (ws_imports),
other imports (etc_imports),
surface water withdrawals for theromelectric power (sw_therm),
surface water withdrawals for potable use (sw_pot),
surface water withdrawals for nonpotable use (sw_npot),
groundwater withdrawls for thermoelectric power (gw_therm),
groundwater withdrawals for potable use (gw_pot),
groundwater withdrawals for nonpotable use (gw_npot),
deep groundwater recharge (dgr),
combined sewer overflow events (cso),
wastewater treatment plant effluent (wtpe)}
}
\value{
list of dataframes for 1) global flows, 2) internal flows, 
         3) storages, 4) global balance, 5) internal balance
balance, flows and 2) storages
}
\description{
This function tracks the environmental and manmade flows gathered using other
CityWaterBalance functions as they move through pathways and storages
within the urban system.
}

