% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineStreamflow.R
\name{combineStreamflow}
\alias{combineStreamflow}
\title{Combines streamflow records to estimate total flows}
\usage{
combineStreamflow(flowlist, mult)
}
\arguments{
\item{flowlist}{list of flow data, output of getStreamflow or 
gapfillStreamflow}

\item{mult}{list of multipliers, one for each gage}
}
\value{
vector of total flow for each timestep
}
\description{
This function combines streamflow data from multiple gages, averaging records
according to a multuplier list
}

