% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfillStreamflow.R
\name{gapfillStreamflow}
\alias{gapfillStreamflow}
\title{Fill gaps in streamflow records}
\usage{
gapfillStreamflow(paired_gages, flowlist)
}
\arguments{
\item{flowlist}{list of gage data and information from getStreamflow}

\item{list}{of gage pairs, where pair=c(gappy,master)}
}
\value{
list list of 2: sites, gap-filled flow data as xts object
}
\description{
This function fills in gappy streamflow records using median of discharge 
ratio with master gages.
}

