% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtmoFlows.R
\name{getAtmoFlows}
\alias{getAtmoFlows}
\title{Gather and consolidate all data time series for water balance}
\usage{
getAtmoFlows(start, end, geometry, attribute)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{geometry}{name of geometry as displayed in Geo Data Portal}

\item{attribute}{name of geometry attribute as displayed in GDP}
}
\value{
data as xts object
}
\description{
This function gathers precipitation and evapotranspiration data, merges these
with streamflow and water use data, calculates the time-varying balance
}

