% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipitation.R
\name{getPrecipitation}
\alias{getPrecipitation}
\title{Assemble a precipitation times series}
\usage{
getPrecipitation(start, end, geometry, attribute)
}
\arguments{
\item{start}{start date in format "YYYY-MM-DD"}

\item{end}{end date in format "YYYY-MM-DD"}

\item{geometry}{"name" of area of interest (AOI) in the GDP}

\item{attribute}{"attribute" of AOI to aggegate over}
}
\value{
prcp
}
\description{
This function aggregates monthly PRISM output for a given area using the USGS 
Geo Data Portal (GDP)
}
\examples{
prcp = getPrecipitation("2010-09-01","2010-12-31",3139,"sample:CONUS_states","STATES")
}

