% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge data sources into input for CityWaterBalance}
\usage{
mergeData(area, pet, inflow, outflow, sfpart, wu, ws_imports = NULL,
  etc_imports = NULL, wweff = NULL, dgr = NULL, cso = NULL)
}
\arguments{
\item{area}{numeric study area (sq km)}

\item{pet}{xts of atmospheric data, from getAtmoFlows (mm/month)}

\item{inflow}{xts  of daily steamflow into area (cfs)}

\item{outflow}{xts of daily streamflow out of area (cfs)}

\item{sfpart}{xts of partitioned streamflow out of area (cfs)}

\item{wu}{xts of water use, from combineWaterUse (MGal/month)}

\item{ws_imports}{xts of imports for water supply (MGal/month)}

\item{etc_imports}{xts of other imports to surface water (MGal/month)}

\item{wweff}{xts of wastwater effluent (cfs)}

\item{dgr}{xts of deep groundwater recharge}

\item{cso}{xts of cso events}
}
\value{
xts of all flows for each timestep
}
\description{
This functions converts units and merges data needed by CityWaterBalance.  All
inputs must span same time period.
}
\examples{
gages = c("05551540","05552500")
flow = getStreamflow("2000-01-01","2010-12-31",gages)
flow = gapfillStreamflow(flow,list(c(gages[1],gages[2])))
flow = combineStreamflow(flow,c(0.5,0.5))

}

