% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionStreamflow.R
\name{partitionStreamflow}
\alias{partitionStreamflow}
\title{Partitions streamflow into essential components}
\usage{
partitionStreamflow(streamflow, inflow = NULL, wastewater = NULL)
}
\arguments{
\item{streamflow}{xts object, e.g. output of combineStreamflow}

\item{inflow}{xts object,}

\item{wastewater}{xts object of wastewater flows}
}
\value{
vector of total flow for each timestep
}
\description{
This functions partitions a streamflow timeseries into baseflow and 
stormflow, with options to separate inflows and wastewater effluent, if data 
are available. All flows should have same units and temporal resolution.
}
\examples{

}
