% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStreamflow.R
\name{plotStreamflow}
\alias{plotStreamflow}
\title{Plot discharge at a set of gages}
\usage{
plotStreamflow(flowlist)
}
\arguments{
\item{flowlist}{list object of flow data output from, e.g., getStreamflow}
}
\value{
plot
}
\description{
This function plots streamflow data
}
