% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWaterBalance.R
\name{plotWaterBalance}
\alias{plotWaterBalance}
\title{Plot urban water balance}
\usage{
plotWaterBalance(data, ys, annual = NULL)
}
\arguments{
\item{data}{xts or zoo object}

\item{ys}{ylim for plot}

\item{annual}{flag indicating whether to plot annual averages}
}
\value{
plot
}
\description{
This function plots components of the urban water balance (selected output 
from getData or CityWaterBalance) showing time-variant flows at monthly or 
annual-aggregated resolution.
}
